+++ libs/minifmod/src/position.h
@@ -1,10 +1,8 @@
 
 #pragma once
 
-#include <cstdint>
-
 struct Position final
 {
-	uint16_t	row;				// current row in pattern
-	uint16_t	order;				// current song order position
+	int row;				// current row in pattern
+	int order;				// current song order position
 };
